code/requirements.txt
pip backports.weakref==1.0.post1
bleach==1.5.0
deep-belief-network==1.0.1
funcsigs==1.0.2
html5lib==0.9999999
Keras==2.0.8
keras-extensions==0.1
Markdown==2.6.10
mock==2.0.0
numpy==1.13.3
pbr==3.1.1
protobuf==3.5.1
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
Werkzeug==0.13
